refactor(activity-bar): extract ActivityBarNarrow (behavior unchanged)
